bug fixed v2.1.0

+++ app/Http/Controllers/SiswaController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Pembayaran;
 use App\Models\Siswa;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
@@ -22,7 +23,6 @@ class SiswaController extends Controller
 
     public function store(Request $request)
     {
-        // Session::get('success');
         $request->validate([
             'nama' => 'required|min:5',
             'kelas' => 'required'
@@ -62,8 +62,15 @@ class SiswaController extends Controller
 
     public function delete($id){
         $siswa = Siswa::find($id);
-        $siswa->delete();
-        // dd($siswa);
-        return redirect('siswa')->with('success', 'Berhasil hapus data!');
+        if ($siswa) {
+            $p = Pembayaran::where('id_siswa', $siswa->id)->exists();
+            // dd($p);
+            if ($p) {
+                return redirect('/siswa')->with('error', 'Data siswa masih dipakai dan tidak dapat dihapus!');
+            }
+
+            $siswa->delete();
+            return redirect('siswa')->with('success', 'Berhasil hapus data!');
+        }
     }
 }
